set up game app
