+++ components/primary-button.jsx
@@ -1,0 +1,16 @@
+import React from "react"
+import { View, Text } from "react-native"
+
+
+// Application Screen || Define Exports
+// =================================================================================================
+// =================================================================================================
+export const PrimaryButton = ({ children }) => {
+  return (
+    <View>
+      <Text>
+        {children}
+      </Text>
+    </View>
+  )
+}
